RX_SAMPLES_TO_FILE: fixing dd timeout
Previously the timeout check loop checked for !joinable, which would
return false while the program was still running. Corrected by changing
it to running.

+++ host/examples/rx_samples_to_file.cpp
@@ -70,19 +70,32 @@ double disk_rate_check(const size_t sample_type_size,
         + " bs=" + std::to_string(samps_per_buff * channel_count * sample_type_size)
         + " count=100";
 
+    bool dd_timeout = false;
     try {
+        // Create child process to run dd disk check
         boost::process::child c(
             disk_check_proc_str, boost::process::std_err > pipe_stream);
 
-        auto timeout = std::chrono::steady_clock::now() + 1s;
-        while(!c.joinable()) {
-            if(timeout > std::chrono::steady_clock::now()) {
-                kill(c.id(), SIGINT);
-            }
+        // Wait for dd to finish with a timeout
+        // NOTE: joinable and running are affected by race conditions when called to soon after starting the child process
+        // wait_for calls join so it doesn't need to be called sperately
+        dd_timeout = !c.wait_for(5s);
+
+        // If dd timed out, kill it and wait for it to finish via join
+        if(dd_timeout) {
+            kill(c.id(), SIGINT);
+            c.join();
         }
-        c.join();
     } catch (std::system_error& err) {
         std::cerr << err_msg << std::endl;
+        if (boost::filesystem::exists(temp_file)) {
+            boost::filesystem::remove(temp_file);
+        }
+        return 0;
+    }
+
+    if(dd_timeout) {
+        std::cerr << "Disk benchmark tool 'dd' timed out" << std::endl;
         if (boost::filesystem::exists(temp_file)) {
             boost::filesystem::remove(temp_file);
         }
